# Conflicts:
#	source/InjectHooksMain.cpp
#	source/game_sa/Entity/Entity.h
#	source/game_sa/Entity/Object/Object.cpp
#	source/game_sa/Rope.cpp

+++ source/game_sa/MenuSystem.cpp
@@ -2,7 +2,7 @@
 
 #include "MenuSystem.h"
 
-uint16& CMenuSystem::num_menus_in_use = *reinterpret_cast<uint16*>(0xBA82E3);
+uint8& CMenuSystem::num_menus_in_use = *reinterpret_cast<uint8*>(0xBA82E3);
 
 // 0x5822D0
 void CMenuSystem::Initialise() {
